refactor: move purchaseCowPen to its own file
