+++ src/game-logic/reducers/purchaseCowPen.js
@@ -1,0 +1,20 @@
+import { moneyTotal } from '../../utils'
+import { PURCHASEABLE_COW_PENS } from '../../constants'
+
+/**
+ * @param {farmhand.state} state
+ * @param {number} cowPenId
+ * @returns {farmhand.state}
+ */
+export const purchaseCowPen = (state, cowPenId) => {
+  const { money, purchasedCowPen } = state
+
+  if (purchasedCowPen >= cowPenId) {
+    return state
+  }
+
+  return {
+    purchasedCowPen: cowPenId,
+    money: moneyTotal(money, -PURCHASEABLE_COW_PENS.get(cowPenId).price),
+  }
+}
