+++ src/game-logic/reducers/purchaseCowPen.test.js
@@ -1,0 +1,20 @@
+import { PURCHASEABLE_COW_PENS } from '../../constants'
+
+import { purchaseCowPen } from './purchaseCowPen'
+
+describe('purchaseCowPen', () => {
+  test('updates purchasedCowPen', () => {
+    const { purchasedCowPen } = purchaseCowPen({}, 1)
+    expect(purchasedCowPen).toEqual(1)
+  })
+
+  test('prevents repurchasing options', () => {
+    const { purchasedCowPen } = purchaseCowPen({ purchasedCowPen: 2 }, 1)
+    expect(purchasedCowPen).toEqual(2)
+  })
+
+  test('deducts money', () => {
+    const { money } = purchaseCowPen({ money: 1500 }, 1)
+    expect(money).toEqual(PURCHASEABLE_COW_PENS.get(1).price - 1500)
+  })
+})
